refactor(store): share nonce and profile validation
