+++ apps/api/internal/store/sqlite/helpers.go
@@ -6,12 +6,10 @@ import (
 	"database/sql"
 	"encoding/json"
 	"errors"
-	"net/url"
 	"regexp"
 	"strings"
 	"sync"
 	"time"
-	"unicode/utf8"
 
 	"github.com/oklog/ulid/v2"
 	"github.com/openclaw/clickclack/apps/api/internal/requestmeta"
@@ -137,20 +135,6 @@ func scanMessage(row scanner) (store.Message, error) {
 	return m, nil
 }
 
-func normalizeClientNonce(value string) (string, error) {
-	nonce := strings.TrimSpace(value)
-	if !utf8.ValidString(nonce) {
-		return "", errors.New("nonce must be valid UTF-8")
-	}
-	if strings.IndexByte(nonce, 0) >= 0 {
-		return "", errors.New("nonce must not contain NUL")
-	}
-	if utf8.RuneCountInString(nonce) > 128 {
-		return "", errors.New("nonce is too long")
-	}
-	return nonce, nil
-}
-
 func getMessageByClientNonceTx(ctx context.Context, tx *sql.Tx, authorID, nonce string) (store.Message, error) {
 	if nonce == "" {
 		return store.Message{}, sql.ErrNoRows
@@ -163,35 +147,6 @@ func sameQuotedMessageID(message store.Message, quotedID string) bool {
 		return message.QuotedMessageID == nil || *message.QuotedMessageID == ""
 	}
 	return message.QuotedMessageID != nil && *message.QuotedMessageID == quotedID
-}
-
-var handlePattern = regexp.MustCompile(`^[a-z0-9][a-z0-9_-]{1,31}$`)
-
-func normalizeHandle(value string) (string, error) {
-	handle := strings.ToLower(strings.TrimSpace(value))
-	handle = strings.TrimPrefix(handle, "@")
-	if handle == "" {
-		return "", nil
-	}
-	if !handlePattern.MatchString(handle) {
-		return "", errors.New("handle must be 2-32 chars using letters, numbers, underscores, or dashes")
-	}
-	return handle, nil
-}
-
-func normalizeAvatarURL(value string) (string, error) {
-	avatarURL := strings.TrimSpace(value)
-	if avatarURL == "" {
-		return "", nil
-	}
-	if len(avatarURL) > 500 {
-		return "", errors.New("avatar_url is too long")
-	}
-	parsed, err := url.Parse(avatarURL)
-	if err != nil || (parsed.Scheme != "https" && parsed.Scheme != "http") || parsed.Host == "" {
-		return "", errors.New("avatar_url must be an http or https URL")
-	}
-	return avatarURL, nil
 }
 
 func resolveProfileAvatarURL(ctx context.Context, q *storedb.Queries, userID, avatarURL string) (string, error) {
